+++ libraries/session/src/main/java/androidx/media3/session/MediaControllerImplLegacy.java
@@ -1604,7 +1604,6 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
             controllerCompat.isSessionReady(),
             controllerCompat.getRatingType(),
             getInstance().getTimeDiffMs(),
-            getRoutingControllerId(controllerCompat),
             hasPendingExtrasChange,
             context);
     Pair<@NullableType Integer, @NullableType Integer> reasons =
@@ -1833,22 +1832,6 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
     listeners.flushEvents();
   }
 
-  @Nullable
-  private static String getRoutingControllerId(MediaControllerCompat controllerCompat) {
-    if (SDK_INT < 30) {
-      return null;
-    }
-    android.media.session.MediaController fwkController =
-        (android.media.session.MediaController) controllerCompat.getMediaController();
-    @Nullable
-    android.media.session.MediaController.PlaybackInfo playbackInfo =
-        fwkController.getPlaybackInfo();
-    if (playbackInfo == null) {
-      return null;
-    }
-    return playbackInfo.getVolumeControlId();
-  }
-
   private static <T> void ignoreFuture(Future<T> unused) {
     // Ignore return value of the future because legacy session cannot get result back.
   }
@@ -2030,7 +2013,6 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
       boolean isSessionReady,
       @RatingCompat.Style int ratingType,
       long timeDiffMs,
-      @Nullable String routingControllerId,
       boolean hasPendingExtrasChange,
       Context context) {
     QueueTimeline currentTimeline;
@@ -2206,8 +2188,7 @@ import org.checkerframework.checker.initialization.qual.UnderInitialization;
     boolean isPlaying =
         LegacyConversions.convertToIsPlaying(newLegacyPlayerInfo.playbackStateCompat);
     DeviceInfo deviceInfo =
-        LegacyConversions.convertToDeviceInfo(
-            newLegacyPlayerInfo.playbackInfoCompat, routingControllerId);
+        LegacyConversions.convertToDeviceInfo(newLegacyPlayerInfo.playbackInfoCompat);
     int deviceVolume =
         LegacyConversions.convertToDeviceVolume(newLegacyPlayerInfo.playbackInfoCompat);
     boolean deviceMuted =